generat next x and y points && edit points type.

+++ planning/src/trajectory/car.hpp
@@ -5,8 +5,8 @@ class Car
 {
 public:
   // all measurements in m && m/s
-  float x{0};
-  float y{0};
+  double x{0};
+  double y{0};
   float yaw{0}; // TODO make sure IMU return it as rad or deg and edit where we use yaw 
   double speed{0};
   float acc{0};
